+++ common/util.ts
@@ -65,9 +65,9 @@ export async function ensureAndLoadSettings(
       settings.indexPage,
       `Hello! And welcome to your brand new SilverBullet space!
 
-<!-- #include [[💭 silverbullet.md/Getting Started]] -->
+<!-- #use [[💭 silverbullet.md/Getting Started]] -->
 Loading some onboarding content for you (but doing so does require a working internet connection)...
-<!-- /include -->`,
+<!-- /use -->`,
     );
   }